Moved existing AddColocatedTaskView into a subfolder
